Move scale function from Caffe function helpers to standard functions.
